feat(worlds): Unified look and feel of file uploaders
