+++ src/components/features/worlds/Locations/OpenLocation.tsx
@@ -5,23 +5,23 @@ import {
   FormControlLabel,
   Grid,
   IconButton,
+  Tooltip,
 } from "@mui/material";
-import BackIcon from "@mui/icons-material/ChevronLeft";
 import DeleteIcon from "@mui/icons-material/Delete";
-import { useLayoutEffect, useRef } from "react";
+import { ChangeEvent, useLayoutEffect, useRef } from "react";
 import { useConfirm } from "material-ui-confirm";
 import { SectionHeading } from "components/shared/SectionHeading";
-import { LocationNameInput } from "./LocationNameInput";
 import { DebouncedOracleInput } from "components/shared/DebouncedOracleInput";
 import { RtcRichTextEditor } from "components/shared/RichTextEditor/RtcRichTextEditor";
 import { LocationDocumentWithGMProperties } from "stores/world/currentWorld/locations/locations.slice.type";
-import { ImageUploader } from "components/features/worlds/ImageUploader";
 import { useStore } from "stores/store";
 import { useListenToCurrentLocation } from "stores/world/currentWorld/locations/useListenToCurrentLocation";
 import { BondsSection } from "components/features/worlds/BondsSection";
 import { LocationNPCs } from "./LocationNPCs";
 import { useWorldPermissions } from "../useWorldPermissions";
 import { ItemHeader } from "../ItemHeader";
+import AddPhotoIcon from "@mui/icons-material/AddPhotoAlternate";
+import { ImageBanner } from "../ImageBanner";
 
 export interface OpenLocationProps {
   worldId: string;
@@ -126,6 +126,16 @@ export function OpenLocation(props: OpenLocationProps) {
       .catch(() => {});
   };
 
+  const fileInputRef = useRef<HTMLInputElement>(null);
+
+  const onFileUpload = (evt: ChangeEvent<HTMLInputElement>) => {
+    const files = evt.target.files;
+    const file = files?.[0];
+    if (file) {
+      uploadLocationImage(locationId, file).catch(() => {});
+    }
+  };
+
   return (
     <Box
       overflow={"auto"}
@@ -133,180 +143,135 @@ export function OpenLocation(props: OpenLocationProps) {
       borderLeft={(theme) => `1px solid ${theme.palette.divider}`}
       width={"100%"}
     >
+      <input
+        type="file"
+        accept={"image/*"}
+        hidden
+        ref={fileInputRef}
+        onChange={onFileUpload}
+      />
       {canShowImages && (
-        <ImageUploader
-          title={location.name}
-          src={location.imageUrl}
-          handleFileUpload={(image) =>
-            uploadLocationImage(locationId, image).catch(() => {})
-          }
-          handleClose={closeLocation}
-        />
+        <ImageBanner title={location.name} src={location.imageUrl} />
       )}
-      <Box
-        sx={(theme) => ({
-          borderTop: `1px solid ${theme.palette.divider}`,
-        })}
-      >
-        <ItemHeader
-          itemName={location.name}
-          updateName={(newName) =>
-            updateLocation(locationId, { name: newName }).catch(() => {})
-          }
-          nameOracleIds={[
-            "ironsworn/oracles/settlement/name/landscape_feature",
-            "ironsworn/oracles/settlement/name/manmade_edifice",
-            "ironsworn/oracles/settlement/name/creature",
-            "ironsworn/oracles/settlement/name/historical_event",
-            "ironsworn/oracles/settlement/name/old_world_language",
-            "ironsworn/oracles/settlement/name/environmental_aspect",
-            [
-              "ironsworn/oracles/settlement/quick_name/prefix",
-              "ironsworn/oracles/settlement/quick_name/suffix",
-            ],
-          ]}
-          actions={
-            <>
-              {showGMFields && (
+      <ItemHeader
+        itemName={location.name}
+        updateName={(newName) =>
+          updateLocation(locationId, { name: newName }).catch(() => {})
+        }
+        nameOracleIds={[
+          "ironsworn/oracles/settlement/name/landscape_feature",
+          "ironsworn/oracles/settlement/name/manmade_edifice",
+          "ironsworn/oracles/settlement/name/creature",
+          "ironsworn/oracles/settlement/name/historical_event",
+          "ironsworn/oracles/settlement/name/old_world_language",
+          "ironsworn/oracles/settlement/name/environmental_aspect",
+          [
+            "ironsworn/oracles/settlement/quick_name/prefix",
+            "ironsworn/oracles/settlement/quick_name/suffix",
+          ],
+        ]}
+        actions={
+          <>
+            {canShowImages && (
+              <Tooltip title={"Upload Image"}>
+                <IconButton onClick={() => fileInputRef?.current?.click()}>
+                  <AddPhotoIcon />
+                </IconButton>
+              </Tooltip>
+            )}
+            {showGMFields && (
+              <Tooltip title={"Delete"}>
                 <IconButton onClick={() => handleLocationDelete()}>
                   <DeleteIcon />
                 </IconButton>
-              )}
-            </>
-          }
-          closeItem={closeLocation}
-        />
-        <Box
-          sx={(theme) => ({
-            mt: 1,
-            px: 2,
-            [theme.breakpoints.up("md")]: { px: 3 },
-          })}
-        >
-          <Grid container spacing={2} sx={{ mb: 2 }}>
-            {showGMFields && (
-              <>
-                {showGMTips && (
-                  <>
-                    <Grid item xs={12}>
-                      <SectionHeading label={"GM Only"} breakContainer />
-                    </Grid>
-                    <Grid item xs={12}>
-                      <Alert severity={"info"}>
-                        Information in this section will not be shared with your
-                        players.
-                      </Alert>
-                    </Grid>
-                  </>
-                )}
-                <Grid item xs={12} md={6}>
-                  <DebouncedOracleInput
-                    label={"Description"}
-                    initialValue={location?.gmProperties?.descriptor ?? ""}
-                    updateValue={(descriptor) =>
-                      updateLocationGMProperties(locationId, {
-                        descriptor,
-                      }).catch(() => {})
-                    }
-                    oracleTableId="ironsworn/oracles/place/descriptor"
-                  />
-                </Grid>
-                <Grid item xs={12} md={6}>
-                  <DebouncedOracleInput
-                    label={"Trouble"}
-                    initialValue={location?.gmProperties?.trouble ?? ""}
-                    updateValue={(trouble) => {
-                      updateLocationGMProperties(locationId, { trouble }).catch(
-                        () => {}
-                      );
-                    }}
-                    oracleTableId={"ironsworn/oracles/settlement/trouble"}
-                  />
-                </Grid>
-                <Grid item xs={12} md={6}>
-                  <DebouncedOracleInput
-                    label={"Location Features"}
-                    initialValue={
-                      location?.gmProperties?.locationFeatures ?? ""
-                    }
-                    updateValue={(locationFeatures) => {
-                      updateLocationGMProperties(locationId, {
-                        locationFeatures,
-                      }).catch(() => {});
-                    }}
-                    oracleTableId={"ironsworn/oracles/place/location"}
-                  />
-                </Grid>
-                {!isSinglePlayer && showGMFields && (
-                  <Grid
-                    item
-                    xs={12}
-                    md={6}
-                    sx={{ alignItems: "center", display: "flex" }}
-                  >
-                    <FormControlLabel
-                      control={
-                        <Checkbox
-                          checked={location.sharedWithPlayers ?? false}
-                          onChange={(evt, value) =>
-                            updateLocation(locationId, {
-                              sharedWithPlayers: value,
-                            }).catch(() => {})
-                          }
-                        />
-                      }
-                      label="Visible to Players"
-                    />
-                  </Grid>
-                )}
-                {isSinglePlayer && (
-                  <BondsSection
-                    isStarforged={false}
-                    hasConnection={false}
-                    onBondToggle={
-                      currentCharacterId
-                        ? (bonded) =>
-                            updateLocationCharacterBond(
-                              locationId,
-                              currentCharacterId,
-                              bonded
-                            ).catch(() => {})
-                        : undefined
-                    }
-                    isBonded={singleplayerBond}
-                    bondedCharacters={bondedCharacterNames}
-                  />
-                )}
-                <Grid item xs={12}>
-                  <RtcRichTextEditor
-                    id={locationId}
-                    roomPrefix={`iron-fellowship-${worldId}-location-gmnotes-`}
-                    documentPassword={worldId}
-                    onSave={updateLocationGMNotes}
-                    initialValue={location.gmProperties?.gmNotes}
-                  />
-                </Grid>
-              </>
+              </Tooltip>
             )}
-            {!isSinglePlayer && (
-              <>
-                {showGMTips && (
-                  <>
-                    <Grid item xs={12}>
-                      <SectionHeading
-                        label={"GM & Player Notes"}
-                        breakContainer
+          </>
+        }
+        closeItem={closeLocation}
+      />
+      <Box
+        sx={(theme) => ({
+          mt: 1,
+          px: 2,
+          [theme.breakpoints.up("md")]: { px: 3 },
+        })}
+      >
+        <Grid container spacing={2} sx={{ mb: 2 }}>
+          {showGMFields && (
+            <>
+              {showGMTips && (
+                <>
+                  <Grid item xs={12}>
+                    <SectionHeading label={"GM Only"} breakContainer />
+                  </Grid>
+                  <Grid item xs={12}>
+                    <Alert severity={"info"}>
+                      Information in this section will not be shared with your
+                      players.
+                    </Alert>
+                  </Grid>
+                </>
+              )}
+              <Grid item xs={12} md={6}>
+                <DebouncedOracleInput
+                  label={"Description"}
+                  initialValue={location?.gmProperties?.descriptor ?? ""}
+                  updateValue={(descriptor) =>
+                    updateLocationGMProperties(locationId, {
+                      descriptor,
+                    }).catch(() => {})
+                  }
+                  oracleTableId="ironsworn/oracles/place/descriptor"
+                />
+              </Grid>
+              <Grid item xs={12} md={6}>
+                <DebouncedOracleInput
+                  label={"Trouble"}
+                  initialValue={location?.gmProperties?.trouble ?? ""}
+                  updateValue={(trouble) => {
+                    updateLocationGMProperties(locationId, { trouble }).catch(
+                      () => {}
+                    );
+                  }}
+                  oracleTableId={"ironsworn/oracles/settlement/trouble"}
+                />
+              </Grid>
+              <Grid item xs={12} md={6}>
+                <DebouncedOracleInput
+                  label={"Location Features"}
+                  initialValue={location?.gmProperties?.locationFeatures ?? ""}
+                  updateValue={(locationFeatures) => {
+                    updateLocationGMProperties(locationId, {
+                      locationFeatures,
+                    }).catch(() => {});
+                  }}
+                  oracleTableId={"ironsworn/oracles/place/location"}
+                />
+              </Grid>
+              {!isSinglePlayer && showGMFields && (
+                <Grid
+                  item
+                  xs={12}
+                  md={6}
+                  sx={{ alignItems: "center", display: "flex" }}
+                >
+                  <FormControlLabel
+                    control={
+                      <Checkbox
+                        checked={location.sharedWithPlayers ?? false}
+                        onChange={(evt, value) =>
+                          updateLocation(locationId, {
+                            sharedWithPlayers: value,
+                          }).catch(() => {})
+                        }
                       />
-                    </Grid>
-                    <Grid item xs={12}>
-                      <Alert severity={"info"}>
-                        Notes in this section will only be visible to gms &
-                        players in campaigns. Notes for singleplayer games
-                        should go in the above section.
-                      </Alert>
-                    </Grid>
-                  </>
-                )}
+                    }
+                    label="Visible to Players"
+                  />
+                </Grid>
+              )}
+              {isSinglePlayer && (
                 <BondsSection
                   isStarforged={false}
                   hasConnection={false}
@@ -323,35 +288,81 @@ export function OpenLocation(props: OpenLocationProps) {
                   isBonded={singleplayerBond}
                   bondedCharacters={bondedCharacterNames}
                 />
-                {!location.sharedWithPlayers && (
+              )}
+              <Grid item xs={12}>
+                <RtcRichTextEditor
+                  id={locationId}
+                  roomPrefix={`iron-fellowship-${worldId}-location-gmnotes-`}
+                  documentPassword={worldId}
+                  onSave={updateLocationGMNotes}
+                  initialValue={location.gmProperties?.gmNotes}
+                />
+              </Grid>
+            </>
+          )}
+          {!isSinglePlayer && (
+            <>
+              {showGMTips && (
+                <>
                   <Grid item xs={12}>
-                    <Alert severity="warning">
-                      These notes are not yet visible to players because this
-                      location is hidden from them.
+                    <SectionHeading
+                      label={"GM & Player Notes"}
+                      breakContainer
+                    />
+                  </Grid>
+                  <Grid item xs={12}>
+                    <Alert severity={"info"}>
+                      Notes in this section will only be visible to gms &
+                      players in campaigns. Notes for singleplayer games should
+                      go in the above section.
                     </Alert>
                   </Grid>
-                )}
+                </>
+              )}
+              <BondsSection
+                isStarforged={false}
+                hasConnection={false}
+                onBondToggle={
+                  currentCharacterId
+                    ? (bonded) =>
+                        updateLocationCharacterBond(
+                          locationId,
+                          currentCharacterId,
+                          bonded
+                        ).catch(() => {})
+                    : undefined
+                }
+                isBonded={singleplayerBond}
+                bondedCharacters={bondedCharacterNames}
+              />
+              {!location.sharedWithPlayers && (
                 <Grid item xs={12}>
-                  {(location.notes || location.notes === null) && (
-                    <RtcRichTextEditor
-                      id={locationId}
-                      roomPrefix={`iron-fellowship-${worldId}-location-`}
-                      documentPassword={worldId}
-                      onSave={updateLocationNotes}
-                      initialValue={location.notes || undefined}
-                    />
-                  )}
+                  <Alert severity="warning">
+                    These notes are not yet visible to players because this
+                    location is hidden from them.
+                  </Alert>
                 </Grid>
-              </>
-            )}
-            <LocationNPCs
-              locationId={locationId}
-              canUseImages={canShowImages ?? false}
-              showHiddenTag={showHiddenTag}
-              openNPCTab={openNPCTab}
-            />
-          </Grid>
-        </Box>
+              )}
+              <Grid item xs={12}>
+                {(location.notes || location.notes === null) && (
+                  <RtcRichTextEditor
+                    id={locationId}
+                    roomPrefix={`iron-fellowship-${worldId}-location-`}
+                    documentPassword={worldId}
+                    onSave={updateLocationNotes}
+                    initialValue={location.notes || undefined}
+                  />
+                )}
+              </Grid>
+            </>
+          )}
+          <LocationNPCs
+            locationId={locationId}
+            canUseImages={canShowImages ?? false}
+            showHiddenTag={showHiddenTag}
+            openNPCTab={openNPCTab}
+          />
+        </Grid>
       </Box>
     </Box>
   );
